Move speedy convection to convection/speedy_convection.py
Update imports in speedy_physics.py, speedy_terms.py, and test file.
