+++ jcm/physics/convection/speedy_convection.py
