+++ jcm/physics/convection/speedy_convection_test.py
@@ -28,7 +28,7 @@ class TestConvectionUnit(unittest.TestCase):
 
         from jcm.physics.speedy.physics_data import ConvectionData, HumidityData, PhysicsData
         from jcm.physics_interface import PhysicsState, PhysicsTendency
-        from jcm.physics.speedy.convection import diagnose_convection, get_convection_tendencies
+        from jcm.physics.convection.speedy_convection import diagnose_convection, get_convection_tendencies
         from jcm.physics.speedy.physical_constants import rgas, cp
         from jcm.physics.speedy.humidity import get_qsat
