+++ jcm/physics/speedy/speedy_physics.py
@@ -60,7 +60,7 @@ class SpeedyPhysics(Physics):
         self.parameters = parameters
 
         from jcm.physics.speedy.humidity import spec_hum_to_rel_hum
-        from jcm.physics.speedy.convection import get_convection_tendencies
+        from jcm.physics.convection.speedy_convection import get_convection_tendencies
         from jcm.physics.speedy.large_scale_condensation import get_large_scale_condensation_tendencies
         from jcm.physics.radiation.speedy_shortwave import get_shortwave_rad_fluxes, get_clouds
         from jcm.physics.radiation.speedy_longwave import get_downward_longwave_rad_fluxes, get_upward_longwave_rad_fluxes
